fix question

+++ src/main/java/org/telegram/winena_bot/scenario/drink_today/jpa/DrinkTodayQuestionRepository.java
@@ -13,7 +13,7 @@ public interface DrinkTodayQuestionRepository extends JpaRepository<DrinkTodayQu
     List<DrinkTodayQuestion> findAllByAuthorIdAndCompleted(int authorId, boolean completed);
     @Query(value = "select d from DrinkTodayQuestion d where d.id not in (:excludedIds)"
             + " order by mod(d.authorId + :seed, mod(:seed, 20)+1), d.id desc")
-    Page<DrinkTodayQuestion> findAllRandomOrder(@Param("seed") long seed, @Param("excludedIds") Collection<Long> excludedIds,
+    Page<DrinkTodayQuestion> findAllRandomOrder(@Param("seed") int seed, @Param("excludedIds") Collection<Long> excludedIds,
                                                 Pageable pageable);
 
 }
